+++ src/client/java/crying/entities/crier/CrierEntityRenderer.java
@@ -34,6 +34,6 @@ public class CrierEntityRenderer extends CrierEntityBaseRenderer<CrierEntity, Cr
 
     @Override
     protected boolean isShaking(CrierEntityRenderState crierEntityRenderState) {
-        return false;
+        return crierEntityRenderState.forlorn;
     }
 }