+++ src/client/java/crying/interfaces/EntityRenderStateVarsInterface.java
@@ -1,0 +1,6 @@
+package crying.interfaces;
+
+public interface EntityRenderStateVarsInterface {
+    int get_realAge();
+    void set_realAge(int age);
+}