+++ src/main/java/crying/entities/GranterEntity.java
@@ -9,6 +9,7 @@ import crying.Crying;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.FlyingItemEntity;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.MovementType;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.entity.data.DataTracker.Builder;
@@ -16,7 +17,6 @@ import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.registry.Registries;
@@ -35,15 +35,15 @@ public class GranterEntity extends Entity implements FlyingItemEntity {
     private int age = 0;
     private boolean descending = false;
 
-    private String targetPlayerUuid;
-    private PlayerEntity targetPlayer;
+    private String targetUuid;
+    private LivingEntity target;
 
     public GranterEntity(EntityType<? extends GranterEntity> type, World world) {
         super(type, world);
         this.noClip = true;
     }
 
-    private static GranterEntity createGranterEntity(World world, PlayerEntity target) {
+    private static GranterEntity createGranterEntity(World world, LivingEntity target) {
         GranterEntity granter = new GranterEntity(Crying.GRANTER_ENTITY, world);
         granter.setTarget(target);
         granter.setPosition(target.getX(), target.getY() + 5d, target.getZ()); 
@@ -52,12 +52,31 @@ public class GranterEntity extends Entity implements FlyingItemEntity {
         return granter;
     }
 
-    public static GranterEntity summonGranterEntity(World world, PlayerEntity target) {
+    private static EvilGranterEntity createEvilGranterEntity(World world, LivingEntity target) {
+        EvilGranterEntity granter = new EvilGranterEntity(Crying.EVIL_GRANTER_ENTITY, world);
+        granter.setTarget(target);
+        granter.setPosition(target.getX(), target.getY() + 5d, target.getZ()); 
+
+        map.put(target.getUuidAsString(), granter);
+        return granter;
+    }
+
+    public static GranterEntity summonGranterEntity(World world, LivingEntity target) {
         if (map.containsKey(target.getUuidAsString()) && map.get(target.getUuidAsString()).getEntityWorld() == target.getEntityWorld()) {
             return null;
         }
 
         GranterEntity entity = createGranterEntity(world, target);
+        world.spawnEntity(entity);
+        return entity;
+    }
+
+    public static EvilGranterEntity summonEvilGranterEntity(World world, LivingEntity target) {
+        if (map.containsKey(target.getUuidAsString()) && map.get(target.getUuidAsString()).getEntityWorld() == target.getEntityWorld()) {
+            return null;
+        }
+
+        EvilGranterEntity entity = createEvilGranterEntity(world, target);
         world.spawnEntity(entity);
         return entity;
     }
@@ -68,25 +87,31 @@ public class GranterEntity extends Entity implements FlyingItemEntity {
 
         World world = this.getEntityWorld();
         if (!world.isClient()) {
-            if (targetPlayer == null && targetPlayerUuid != null) {
-                targetPlayer = world.getPlayerByUuid(UUID.fromString(targetPlayerUuid));
+            if (target == null && targetUuid != null) {
+                try {
+                    target = ((LivingEntity) world.getEntity(UUID.fromString(targetUuid)));
+                }
+                catch(Exception e) {
+
+                }
             }
-            if (targetPlayer == null)
-                targetPlayer = world.getClosestPlayer(this, 64);
-            if (targetPlayer == null || targetPlayer.isDead())
+            if (target == null)
+                target = world.getClosestPlayer(this, 64);
+            if (target == null || target.isDead()) {
+                this.remove(RemovalReason.UNLOADED_WITH_PLAYER);
                 return;
+            }
             else
-                targetPlayerUuid = targetPlayer.getUuidAsString();
+                targetUuid = target.getUuidAsString();
 
             age++;
             if (!descending) {
                 if (age < FOLLOW_TICKS) {
                     double bob = Math.sin(this.age * 0.2) * 0.25;
-                    Vec3d targetPos = targetPlayer.getEntityPos().add(0, 3d + bob, 0);
+                    Vec3d targetEntityPos = target.getEntityPos().add(0, 3d + bob, 0);
 
-                    world.getClosestPlayer(targetPlayer, bob);
                     double speed = 0.4d;
-                    Vec3d newPos = this.getEntityPos().lerp(targetPos, speed);
+                    Vec3d newPos = this.getEntityPos().lerp(targetEntityPos, speed);
                     this.setPosition(newPos);
                 } 
                 else {
@@ -96,14 +121,17 @@ public class GranterEntity extends Entity implements FlyingItemEntity {
             } 
             else {
                 Vec3d currentPos = this.getEntityPos();
-                Vec3d targetPos = targetPlayer.getEntityPos().add(0, 1.0, 0);
-                Vec3d direction = targetPos.subtract(currentPos).normalize().multiply(0.2d);
+                Vec3d targetEntityPos = target.getEntityPos().add(0, 1.0, 0);
+                Vec3d direction = targetEntityPos.subtract(currentPos).normalize().multiply(0.2d);
 
                 this.setVelocity(direction);
                 this.move(MovementType.SELF, this.getVelocity());
 
-                if (this.getBoundingBox().intersects(targetPlayer.getBoundingBox())) {
+                if (this.getBoundingBox().intersects(target.getBoundingBox())) {
                     this.remove(RemovalReason.KILLED);
+                }
+                else if (target.isDead()) {
+                    this.remove(RemovalReason.UNLOADED_WITH_PLAYER);
                 }
             }
         }
@@ -111,30 +139,35 @@ public class GranterEntity extends Entity implements FlyingItemEntity {
 
     @Override
     public void remove(RemovalReason reason) {
-        if (map.containsKey(targetPlayerUuid))
-            map.remove(targetPlayerUuid);
+        if (map.containsKey(targetUuid))
+            map.remove(targetUuid);
         
         if (reason == RemovalReason.KILLED) {
             ArrayList<StatusEffect> effects = new ArrayList<>();
             StatusEffect effect = null;
             for (StatusEffect e : Registries.STATUS_EFFECT) {
-                if (e.getCategory() == StatusEffectCategory.BENEFICIAL) {
+                if (e.getCategory() == (isEvil() ? StatusEffectCategory.HARMFUL : StatusEffectCategory.BENEFICIAL)) {
                     effects.add(e);
                 }
             }
-            effects.addAll(Collections.nCopies(3, StatusEffects.REGENERATION.value()));
-            effects.addAll(Collections.nCopies(3, StatusEffects.SATURATION.value()));
-            effects.addAll(Collections.nCopies(3, StatusEffects.INSTANT_HEALTH.value()));
+            if (!isEvil()) {
+                effects.addAll(Collections.nCopies(3, StatusEffects.REGENERATION.value()));
+                effects.addAll(Collections.nCopies(3, StatusEffects.SATURATION.value()));
+                effects.addAll(Collections.nCopies(3, StatusEffects.INSTANT_HEALTH.value()));
+            }
 
             Random random = this.getEntityWorld().getRandom();
             while (effect == null) {
                 effect = effects.get(random.nextInt(effects.size()));
             }
 
-            this.getEntityWorld().addParticleClient(ParticleTypes.GLOW, targetPlayer.getX(), targetPlayer.getY(), targetPlayer.getZ(), 1d, 1d, 1d);
-            this.targetPlayer.addStatusEffect(new StatusEffectInstance(Registries.STATUS_EFFECT.getEntry(effect), random.nextBetween(5, 30) * 20, random.nextBetween(0, 1)));
-            this.targetPlayer.heal(2F);
-            this.getEntityWorld().playSound(this, BlockPos.ofFloored(this.getEntityPos().x, this.getEntityPos().y, this.getEntityPos().z), Crying.GRANTER_HEAL_EVENT, SoundCategory.PLAYERS);
+            this.getEntityWorld().addParticleClient(ParticleTypes.GLOW, target.getX(), target.getY(), target.getZ(), 1d, 1d, 1d);
+            this.target.addStatusEffect(new StatusEffectInstance(Registries.STATUS_EFFECT.getEntry(effect), random.nextBetween(5, 30) * 20, random.nextBetween(0, 1)));
+            if (!isEvil()) {
+                this.target.heal(2F);
+            }
+
+            this.getEntityWorld().playSound(this, BlockPos.ofFloored(this.getEntityPos().x, this.getEntityPos().y, this.getEntityPos().z), isEvil() ? Crying.EVIL_GRANTER_HEAL_TOUCH : Crying.GRANTER_HEAL_EVENT, SoundCategory.AMBIENT);
         }
 
         super.remove(reason);
@@ -143,18 +176,18 @@ public class GranterEntity extends Entity implements FlyingItemEntity {
     @Override
     protected void readCustomData(ReadView nbt) {
         age = nbt.getInt("age", 0);
-        targetPlayerUuid = nbt.getString("targetPlayerUuid", null);
+        targetUuid = nbt.getString("targetUuid", null);
         descending = nbt.getBoolean("descending", false);
-        if (targetPlayerUuid != null)
-            map.put(targetPlayerUuid, this);
+        if (targetUuid != null)
+            map.put(targetUuid, this);
     }
 
     @Override
     protected void writeCustomData(WriteView nbt) {
         nbt.putInt("age", age);
         nbt.putBoolean("descending", descending);
-        if (targetPlayerUuid != null)
-            nbt.putString("targetPlayerUuid", targetPlayerUuid);
+        if (targetUuid != null)
+            nbt.putString("targetUuid", targetUuid);
     }
 
     @Override
@@ -171,15 +204,19 @@ public class GranterEntity extends Entity implements FlyingItemEntity {
         return new ItemStack(Crying.GRANTER);
     }
 
-    public String getTargetPlayerUuid() {
-        return this.targetPlayerUuid;
+    public String gettargetUuid() {
+        return this.targetUuid;
     }
 
-    public void setTarget(PlayerEntity entity) {
+    public void setTarget(LivingEntity entity) {
         if (entity == null)
             return;
 
-        this.targetPlayer = entity;
-        this.targetPlayerUuid = entity.getUuidAsString();
+        this.target = entity;
+        this.targetUuid = entity.getUuidAsString();
+    }
+
+    public final boolean isEvil() {
+        return this instanceof EvilGranterEntity;
     }
 }